Move into pkg/data instead
