+++ pkg/data/components/EmaThroughput.yaml
