+++ pkg/data/components/FilterProcessor.yaml
