+++ pkg/data/components/OTelDebugExporter.yaml
