+++ pkg/data/components/OTelGRPCExporter.yaml
