+++ pkg/data/components/OTelHTTPExporter.yaml
