+++ pkg/data/components/OTelReceiver.yaml
