+++ pkg/data/loadEmbedded_test.go
@@ -1,52 +1,53 @@
-package config
+package data
 
 import (
 	"os"
 	"path"
 	"testing"
 
+	"github.com/honeycombio/hpsf/pkg/config"
 	"github.com/stretchr/testify/require"
 )
 
-func TestLoadTemplateComponents(t *testing.T) {
-	got, err := LoadTemplateComponents()
+func TestLoadLocalComponents(t *testing.T) {
+	got, err := LoadEmbeddedComponents()
 	if err != nil {
-		t.Errorf("LoadTemplateComponents() error = '%v', want nil", err)
+		t.Errorf("LoadEmbeddedComponents() error = '%v', want nil", err)
 	}
 	if len(got) == 0 {
-		t.Errorf("LoadTemplateComponents() = %v, want non-empty", got)
+		t.Errorf("LoadEmbeddedComponents() = %v, want non-empty", got)
 	}
 }
 
 func TestTemplateComponents(t *testing.T) {
-	components, err := LoadTemplateComponents()
+	components, err := LoadEmbeddedComponents()
 	require.NoError(t, err)
 	// for test component type
 	tests := []struct {
 		name       string
 		kind       string
-		cType      Type
+		cType      config.Type
 		config     map[string]any
 		wantOutput string
 	}{
 		{
 			name:       "HoneycombExporter to refinery config",
 			kind:       "HoneycombExporter",
-			cType:      RefineryConfigType,
+			cType:      config.RefineryConfigType,
 			config:     map[string]any{"APIKey": "test"},
 			wantOutput: "HoneycombExporter_output_refinery_config.yaml",
 		},
 		{
 			name:       "DeterministicSampler to refinery rules",
 			kind:       "DeterministicSampler",
-			cType:      RefineryRulesType,
+			cType:      config.RefineryRulesType,
 			config:     map[string]any{"Environment": "staging", "SampleRate": 42},
 			wantOutput: "DeterministicSampler_output_refinery_rules.yaml",
 		},
 		{
 			name:  "EMAThroughputSampler to refinery rules",
 			kind:  "EMAThroughput",
-			cType: RefineryRulesType,
+			cType: config.RefineryRulesType,
 			config: map[string]any{
 				"Environment":        "staging",
 				"GoalThroughput":     42,
