+++ pkg/data/templates/emathroughput.yaml
