+++ pkg/data/templates/proxy.yaml
