+++ pkg/data/testdata/DeterministicSampler_output_refinery_rules.yaml
