+++ pkg/data/testdata/EmaThroughput_output_refinery_rules.yaml
